soundwire: intel: split device and shim parts
The device handling is rather generic, while the shim parts are tied
to a specific hardware version. Split the two parts to allow for reuse
of generic part and abstraction of the hardware behind hardware
specific ops in the future.
