+++ drivers/soundwire/intel_init.c
@@ -17,6 +17,7 @@
 #include <linux/soundwire/sdw_intel.h>
 #include "cadence_master.h"
 #include "intel.h"
+#include "intel-device.h"
 
 static void intel_link_dev_release(struct device *dev)
 {
